HPCC-17101 Initial version
The JAPI wsclient is used to get file information from DALI.

Remote read is used by the RDD to access the file parts
in parallel.

In addition to a normal RDD<Record> (where Record is a
generic container for an HPCC record), a RDD<Vector>
and RDD<LabeledPoint> are provided for use with the
classic Spark API.  A Dataset<Row> is also provided
for use with the new Dataframe API.

The Row is implemented with an HpccRow object built from
a Record object.

Signed-off-by: johnholt <john.d.holt@lexisnexisrisk.com>