HPCC-19375 Index pull into Spark environment
Reads index datasets with filtering.  Increased the number of datatypes supported and suppressed datatypes that should not be supported.  Indexes have an additional file part, so the relationship between Spark partitions and file parts was made indirect.

ColumnPruner was enhanced to suppress datatypes that are unsupported and to change the output format of particular types (like QSTRING) into a more standard type.  The enhancement eliminated shortcut processing through the ColumnPruner.

HpccDataFrameFactory was modified to use the replacement of FilePart with HpccPart.

HpccFile was enhanced by adding several convenience constructors and by adding the display of the JSON layout definition in the event of a problem processing the definition.  The FilePart object was refactored into an HpccPart object employed by HpccFile and a DataPart object.

HpccPart is the result of refactoring FilePart to make the relationship between Spark partitions and Thor file partitions indirect.

HpccRDD was modified to use the replacement of FilePart with HpccPart.

HpccRemoteFileReader was modified to use the new HpccPart.

package-info updated to reflect the change in class name of FilePart to HpccPart.

AddrRemapper was enhanced to work in cases where there are more file parts than nodes.  Cleaned up the implementation to get a list of the unique IP values instead of assuming that the IP values were unique.

BinaryRecordReader was modified to use the DataPart object that was refactored from the FilePart.

ClusterRemapper was modified to remove the number of parts.

DataPartition was created by refactoring FilePart.

DefEntry removed unused constructor and added tracking of type, length, and type flag.  The short cut processing for the no pruning case was removed.

DefEntryField was modified to track if the definition should be suppressed; or if the definition had been modified.  An unused empty constructor was removed.

DefEntryRoot was modified to remove an unused constructor and to track suppressions and type changes.

DefEntryType was modified to remove an unused constructor and to track suppression and type changes.

DefToken was enhanced by adding comments and a new constructor used for the update of type and length information.

FieldFilter class was added to capture filter criteria for a specified field name.  the constructor takes FieldFilterRanges.

FieldFilterRange class capture the properties for a filter range.

FileFilter class was added to capture filter information from either an array of FieldFilter objects or a string version of the filter expression.

PlainConnection updated for the FilePart refactor.  Updated to request an index read instead of a disk read when the file is an index, and updated to add the field filter to the read request.

RemapInfo was updated to use zero nodes for the empty constructor.

TypeDef was enhanced to support type revision and suppression.

DataframeTest was updated for the FilePart refactor.

HpccFiletest was updated for the FilePart refactor and to add file filtering and to allow the user to specify which part to read.

RDDTest was updated for the FilePart refactor.

REcordTest was updated to support Field filters and the FilePart refactor.